% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affairs}
\alias{affairs}
\title{Affairs dataset}
\format{
A (tibble) data frame with 601 rows and 9 variables.
}
\source{
\url{https://www.cengage.com/cgi-wadsworth/course_products_wp.pl?fid=M20b&product_isbn_issn=9781111531041}
}
\usage{
data(affairs)
}
\description{
Data on extramarital affairs from R.C. Fair (1978), "A Theory of Extramarital Affairs," Journal of Political Economy 86, 45-61. Originally obtained from a survey by Psychology Today.
The variables are as follows:
}
\details{
\itemize{
\item male. =1 if male.
\item age. Age in years.
\item years_married. Years married.
\item kids. =1 if have kids.
\item relig. Religiosity (5 = very religious, 4 = somewhat, 3 = slightly, 2 = not at all, 1 = anti).
\item educ. Years of schooling.
\item occup. Occupation, reverse Hollingshead scale.
\item marriage_rating. Marriage rating (5 = very happy, 4 = happier than avg, 3 = avg, 2 = somewhat unhappy, 1 = very unhappy).
\item n_affairs. Number of affairs within last year.
}
}
\keyword{datasets}
