% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{art}
\alias{art}
\title{Art sales}
\format{
A (tibble) data frame with 18,400 rows and 13 variables.
}
\source{
Lee, K., Park, J., Goree, S. et al. Social signals predict contemporary art prices better than visual features, particularly in emerging markets. Sci Rep 14, 11615 (2024).
}
\usage{
data(art)
}
\description{
dataset of auction sales records, including artists’ attributes, and market information.
The variables are as follows:
}
\details{
\itemize{
\item case_id. Unique identifier for each auction sale.
\item artist_name. Name of the artist.
\item lot_title. Title of the artwork.
\item year_of_artwork. Year the artwork was created.
\item event_year. Year the auction took place.
\item nationality. Artist nationality.
\item year_of_birth. Artist year of birth.
\item genre. Genre of the artwork.
\item country. Country where the auction took place.
\item real_price_usd. Price of the artwork in real USD.
\item size_inch_by_inch. Size of the artwork in inches.
\item gender_male. Dummy variable for whether artist is male.
\item elite_school. Dummy variable for whether artist attended an elite school.
}
}
\keyword{datasets}
