% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big_five}
\alias{big_five}
\title{Data on the Big Five personality traits}
\format{
A (tibble) data frame with 307,313 rows and 8 variables.
}
\source{
For info on the traits see here: https://en.wikipedia.org/wiki/Big_Five_personality_traits#Descriptions_of_the_particular_personality_traits. Data comes from here: https://github.com/automoto/big-five-data
}
\usage{
data(big_five)
}
\description{
Dataset that provides big five personality trait scores for 307,313 people from different countries.
The variables are as follows:
}
\details{
\itemize{
\item country. Country where respondent lives.
\item age. Age of respondent.
\item sex. Sex of respondent.
\item agreeable_score. Agreeableness score, ranging from 0 (critical / rational) to 1 (friendly/compassionate).
\item extraversion_score. Extraversion score, ranging from 0 (solitary/reserved) to 1 (outgoing/energetic).
\item openness_score. Openness score, ranging from 0 (consistent/cautious) to 1 (inventive/curious).
\item conscientiousness_score. Consccientiousness score, ranging from 0 (extravagant/careless) to 1 (efficient/organized).
\item neuroticism_score. Neuroticism score, ranging from 0 (resilient/confident) to 1 (sensitive/nervous).
}
}
\keyword{datasets}
