% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bonica}
\alias{bonica}
\title{Data on Campaign Finance from the Database on Ideology, Money in Politics, and Elections (DIME).}
\format{
A (tibble) data frame with 9,787 rows and 13 variables.
}
\source{
The Database on Ideology, Money in Politics, and Elections (DIME).
}
\usage{
data(bonica)
}
\description{
Candidate level data on money raised, spent, and other stuff during political campaigns, from the Database on Ideology, Money in Politics, and Elections (DIME).
The variables are as follows:
\itemize{
\item cycle. The election cycle/year, includes: 2012, 2016, 2018.
\item name. Name of candidate.
\item party. Party of candidate.
\item dwnom1. First dimension common-space DW-NOMINATE score. Based on joint scaling of the 1st to the 112th Congresses. More negative = more liberal, more positive = more conservative.
\item state. State.
\item district. District identifier.
\item seat. Senate or House election.
\item incumbent. Whether candidate is an incumbent or a challenger.
\item gender. Gender of candidate.
\item num_distinct_donors. Number of district donors that gave to the candidate in the election cycle.
\item total_receipts. Total dollars raised by candidate during an election cycle.
\item contribs_from_candidate. Total receipts from candidate contributions.
\item total_disbursements. Total campaign disbursements (in dollars) for the given election cycle.
}
}
\keyword{datasets}
