% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bot}
\alias{bot}
\title{American Voter Bot data}
\format{
A (tibble) data frame with 61,000 rows and 19 variables.
}
\source{
2020 Cooperative Election Survey.
}
\usage{
data(bot)
}
\description{
A sample of respondents from the Cooperative Election Survey. Inspired by the Twitter account "American Voter Bot".
The variables are as follows:
\itemize{
\item state. Respondent home state.
\item sex. Respondent sex.
\item age. Respondent age.
\item educ. Respondent highest educational attainment.
\item race. Respondent race.
\item pid7. Respondent ideological self-placement (7-point scale).
\item ideo5. Respondent ideological self-placement (5-point scale).
\item religion. Respondent religion.
\item votechoice. Respondent vote choice during 2020 election.
\item hispanic. Is respondent Hispanic?
\item know_governor. Respondent answer to which party they believe their current governor belongs to. A measure of political knowledge.
\item conceal. Support or oppose: "Make it easier for people to obtain concealed carry permits?"
\item prochoice. Support or oppose: "Always allow a woman to obtain an abortion as a matter of choice."
\item cleanair. Support or oppose: "Strengthen the Environmental Protection Agency enforcement of the Clean Air Act and Clean Water Act even if it costs U.S. jobs".
\item wall. Support or oppose: "Increase spending on border security by $25 billion, including building a wall between the U.S. and Mexico."
\item mandmin. Support or oppose: "Eliminate mandatory minimum sentences for non- violent drug offenders."
\item aca. Support or oppose: "Repeal the entire Affordable Care Act."
\item minwage. Support or oppose: "Raise the minimum wage to $15 an hour."
\item newsint. How often Respondent says they follow what's going on in government and public affairs.
}
}
\keyword{datasets}
