% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census}
\alias{census}
\title{Immigration data from Twelfth Census of the United States (1900)}
\format{
A (tibble) data frame with 161 rows and 49 variables.
}
\source{
Stacker media: \url{https://github.com/stacker-media/data/tree/main/1900-census-immigrant-city}
}
\usage{
data(census)
}
\description{
This dataset is a transcribed version of the data from Table 35 of the Twelfth Census of the United States, completed in 1900. The data provide information on the foreign born in major metro areas around the US. Data was transcribed by Stacker Media.
The variables are as follows:
}
\details{
\itemize{
\item city. City name.
\item state. Name of state.
\item total_foreign_born. total foreign born population in the city (1900).
\item columns 3-49. Foreign born population from the indicated group.
}
}
\keyword{datasets}
