% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childcare}
\alias{childcare}
\title{National Database of Childcare Prices}
\format{
A (tibble) data frame with 34,567 rows and 15 variables.
}
\source{
National Database of Childcare Prices from the Department of Labor. https://www.dol.gov/agencies/wb/topics/featured-childcare.
}
\usage{
data(childcare)
}
\description{
A dataset that provides information on childcare prices and other demographics at the county-year level in the United States. Childcare prices are based on market rate surveys.
The variables are as follows:
}
\details{
\itemize{
\item state. State.
\item county. County.
\item fips. Five-digit number that uniquely identifies the county in a state.
\item year. Year the data collection began for the market rate survey.
\item unem_rate. Unemployment rate of the population aged 16 years old or older (%).
\item poverty_rate. Poverty rate for families.
\item median_hh_income. Median household income.
\item pop. Count of the total population.
\item black. Percent of population that identifies as being one race and being only Black or African American.
\item hispanic. Percent of population that identifies as being Hispanic or Latino regardless of race.
\item white. Percent of population that identifies as being one race and being only White or Caucasian.
\item infant_price. Aggregated weekly, full-time median price charged for Center-based Care for infants (i.e. aged 0 through 23 months).
\item toddler_price. Aggregated weekly, full-time median price charged for Center-based Care for toddlers (i.e. aged 24 through 35 months).
\item preschool_price. Aggregated weekly, full-time median price charged for Center-based Care for preschoolers (i.e. aged 36 through 54 months).
\item infant_family_price. Aggregated weekly, full-time median price charged for Family Childcare for infants (i.e. aged 0 through 23 months).
}
}
\keyword{datasets}
