% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colony}
\alias{colony}
\title{Data on Sale of Provincial Governorships in Peru, 17th-18th century}
\format{
A (tibble) data frame with 632 rows and 36 variables.
}
\source{
Guardado, J. (2018). Office-selling, corruption, and long-term development in Peru. American Political Science Review, 112(4), 971-995.
}
\usage{
data(colony)
}
\description{
A cross-sectional dataset containing information on the sale of provincial governorships by the Spanish Crown in #' Peru during the 17th and 18th century, as well as province-level demographic and economic data. The dataset is  #' from Guardado (2018). The variables are as follows:
}
\details{
\itemize{
\item audiencia. Spanish Audiencia.
\item provincia. colonial province name.
\item provcode. province code.
\item year. year of sale.
\item time. sequence in sale.
\item name_approx. name of purchaser (approximate).
\item appointed. appointed (not sold)? 1=yes.
\item title_raw. social status (as described in Crown's docs).
\item military. military career (or rank)? 1=yes.
\item orden. holds a knight title (caballero)? 1=yes.
\item nobility. holds other nobility titles? 1=yes.
\item noble. holds either a knight or other nobility titles? 1=yes.
\item rprice1. real office prices (in pesos) 1 peso=272 maravedies.
\item suitindex. avg agricultural suitability in province (1=max suitability).
\item centerxgd. avg longitude (based on center of districts in province).
\item centerygd. avg latitude (based on center of districts in province).
\item z. avg elevation.
\item distlima. avg distance to Lima (based on center of districts in province).
\item bishop. is the province a bishop seat? 1=yes.
\item cumwar. cumulative years of Spanish involvement in European wars.
\item war. indicator for years of Spanish involvement in European wars.
\item twowars. cumulative years of Spanish involvement in succession wars.
\item yearfromwar. cumulative years into / from the following European war.
\item pop54. province population levels in 1754 (Golte).
\item ind54. province indigenous population levels in 1754 (Golte).
\item gov_reb. anti-governor rebellion in year of purchase? 1=yes.
\item mita. province to provice forced labor to mines? 1=yes.
\item mine. gold or silver mine in province? 1=yes.
\item econactivity2. province is a market hub (circuito comercial)? 1=yes.
\item wage. wages (in pesos) during office-selling period.
\item reparto2. assigned repartimiento quota (in pesos) in 1754.
\item rep50. above the median of repartimiento quotas? 1=yes.
\item totalc. log total revenue at caja level.
\item tributonew. log total revenue from head taxes (tributo).
\item mining. log total revenue from mining taxes.
\item alcabala. log total revenue from sale taxes (alcabala).
}
}
\keyword{datasets}
