% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colony_long}
\alias{colony_long}
\title{Data on Sale of Provincial Governorships in Peru, 17th-18th century, and Contemporary Outcomes}
\format{
A (tibble) data frame with 14,494 rows and 90 variables.
}
\source{
Guardado, J. (2018). Office-selling, corruption, and long-term development in Peru. American Political Science Review, 112(4), 971-995.
}
\usage{
data(colony_long)
}
\description{
A dataset linking information on the sale of provincial governorships by the Spanish Crown in Peru during the 17th and 18th century with contemporary development outcomes. The dataset is from Guardado (2018).
The variables are as follows:
}
\details{
\itemize{
\item provincia. colonial province name.
\item provcode. province code.
\item district. contemporary district name.
\item ubigeo. contemporary district code enaho 2013.
\item conglome. conglomerate code enaho 2013.
\item vivienda. house code enaho 2013.
\item hogar. household code enaho 2013.
\item codperso. individual code enaho 2013.
\item lhhequivlm. log of equivalent household consumption minus transfers in Lima prices.
\item lhhequiv. log of equivalent household consumption minus transfers in normal prices.
\item schoolyears. highest educational level attained (0-21 years).
\item QUE. Language learned as a child: Quechua? 1=yes?.
\item male. respondent is male? 1=yes.
\item age. respondent age.
\item kids. # household members aged 0 to 14.
\item infants. # household members aged 0 to 4.
\item adults. # adult household members.
\item localborn. respondent was born in current district where living? 1=yes.
\item trust_jne. Trust in Jurado Nacional de Eleccions? 1=a lot/ sufficient.
\item trust_onpe. Trust in Oficina Nacional de Procesos Electorales? 1=a lot/sufficient.
\item trust_munic. Trust in provincial government? 1=a lot/sufficient.
\item trust_distrit. Trust in district government? 1=a lot/sufficient.
\item trust_police. Trust in national police? 1=a lot/sufficient.
\item trust_army. Trust in national army? 1=a lot/sufficient.
\item trust_judic. Trust in judicial branch? 1=a lot/sufficient.
\item trust_government. Trust in regional government? 1=a lot/sufficient.
\item trust_paper. Trust in written media? 1=a lot/sufficient.
\item trust_radtv. Trust in radio or TV? 1=a lot/sufficient.
\item peasant_id. Identified with indigenous peasant groups? 1=yes.
\item pop1572. district population in 1572.
\item logtributario. log head tax paying individuals in district 1572.
\item ltrib_rate. log tax rate in 1572.
\item perctribut. share population tax paying.
\item percviejos. share population elder.
\item percmuchachos. share population young.
\item percmujeres. share population women.
\item shpriest. share of budget for priests.
\item shcorreg. share of budget for colonial administration.
\item shcaciq. share of budget for indigenous rulers.
\item shencom. share of budget for encomendero.
\item toilethouse. share district households with indoor toilets.
\item mud. share district households with mud wall material.
\item shindig. share indigenous population in district 1876.
\item shnonlit. share of illiterate population in district 1876.
\item shindig80. share indigenous population in district 1780.
\item suitindex. avg agricultural suitability in district.
\item z. elevation.
\item centerxgd. longitude (based on center of district).
\item centerygd. latitude (based on center of district).
\item distlima. distance to Lima (based on center of district).
\item lz. log elevation.
\item ldistlima. log avg distance to Lima (based on center of district).
\item pop1827. log province population 1827.
\item grossinc1827. gross provincial income 1827.
\item pcinc1827. per capital provincial income 1827.
\item wage. wages (in pesos) throughout period.
\item mita. province to provice forced labor to mines? 1=yes.
\item mine. gold or silver mine in province? 1=yes.
\item lreparto2. log assigned repartimiento quota (in pesos) in 1754.
\item pop54. province population levels in 1754 (Golte).
\item ind54. province indigenous population levels in 1754 (Golte).
\item tribut. number of tax paying indigenous in province in 1754.
\item orig. number of tax paying indigenous born in province in 1754.
\item forastero. number of tax paying indigenous born outside province in 1754.
\item lpop54. log province population levels in 1754 (Golte).
\item lind54. log province indigenous population levels in 1754 (Golte).
\item postsales. year of known first appointment post office-selling period.
\item econactivity2. province is a market hub (circuito comercial)? 1=yes.
\item bishop. is the province a bishop seat? 1=yes.
\item lwage. log wages during office-selling period.
\item pop90. district population in 1990.
\item lpop90. log district population in 1990.
\item allpositions. total governorships appointed or sold during office-selling.
\item sold. total governorships sold in province during office-selling.
\item meanprice. avg log office price per capita during war.
\item meanpriceh. avg log office price per capita during peace.
\item minpriceh. minimum log office price per capita during peace.
\item firstpriceh. first log office price per capita during peace.
\item border. district bordering other provinces.
\item rep50. above the median of repartimiento quota? 1=yes.
\item soldpc. share of sold positions by province, weighted by population.
\item cases83. All violent incidents in district 1980-83.
\item authority83. All violent attacks against political authorities 1980-1983.
\item guerrilla83. All violent attacks by Shining Path 1980-83.
\item anyreb_pre. total rebellions during office-selling (1708-1752).
\item govreb_pre. total anti-governor rebellions during office-selling (1708-1752).
\item alltax_pre. total tax-related rebellions during office-selling (1708-1752).
\item anyreb_post. total rebellions post office-selling (1752-1780).
\item govreb_post. total anti-governor rebellions post office-selling (1752-1780).
\item alltax_post. total tax-related rebellions post office-selling (1752-1780).
}
}
\keyword{datasets}
