% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conflict}
\alias{conflict}
\title{Conflict onset}
\format{
A (tibble) data frame with 11,493 rows and 11 variables.
}
\source{
From the `peacesciencer` package.
}
\usage{
data(conflict)
}
\description{
State-year level UCDP data focusing on conflict onset.
The variables are as follows:
\itemize{
\item country. Name of country.
\item year. Year.
\item onset. Did country experience at least one conflict onset this year? (1 = yes, 0 = no).
\item imports. Total imports of the state in current million USD.
\item exports. Total exports of the state in current million USD.
\item polyarchy. the Varieties of Democracy "polyarchy" estimate of a country's quality of electoral democracy.
\item milex. An estimate of military expenditures (in thousands).
\item milper. An estimate of the size of military personnel (in thousands) for the state.
\item The Composite Index of National Capability ("CINC") score. Roughly: a country's world share of military capability in that year.
\item gdp. Estimated GDP in 2011 USD.
\item pop. Estimated population size.
}
}
\keyword{datasets}
