% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coups}
\alias{coups}
\title{Cline Center Coup d'État Project Dataset}
\format{
A (tibble) data frame with 981 rows and 22 variables.
}
\source{
Peyton, Buddy, Joseph Bajjalieh, Dan Shalmon, Michael Martin, Jonathan Bonaguro, and Emilio Soto. 2023. Cline Center Coup d'État Project Dataset. Cline Center for Advanced Social Research. V.2.1.2. February 23. University of Illinois Urbana-Champaign.
}
\usage{
data(coups)
}
\description{
This dataset from the Cline Center provides information on all known failed and successful coup events for every country in the world.
The variables are as follows:
}
\details{
\itemize{
\item country. Name of the country where coup event occurred.
\item year. Year of event.
\item month. Month of event.
\item day. Day of event.
\item event_type. Indicates whether the event is a coup, attempted coup, or conspiracy.
\item unrealized. A dummy variable where one indicates an unsuccessful coup or plot and zero otherwise.
\item realized. A dummy variable where one indicates a successful coup and zero otherwise.
\item conspiracy. A dummy variable where one indicates a coup conspiracy thwarted prior to execution and zero otherwise.
\item attempt. A dummy variable where one indicates a coup was attempted by failed and zero otherwise.
\item military. A dummy variable where one indicates a military coup/attempt/conspiracy and zero otherwise.
\item foreign. A dummy variable where one indicates a foreign-backed coup/attempt/conspiracy and zero otherwise.
\item auto. A dummy variable where one indicates an auto coup and zero otherwise.
\item popular. A dummy variable where one indicates a popular revolt and zero otherwise.
\item counter. A dummy variable where one indicates a counter-coup and zero otherwise.
\item noharm. A dummy variable where one indicates the deposed executive was not injured or killed during the coup event and zero otherwise.
\item injured. A dummy variable where one indicates the deposed executive was injured during the coup event and zero otherwise.
\item killed. A dummy variable where one indicates the deposed executive was killed during the coup event and zero otherwise.
\item harrest. A dummy variable where one indicates the deposed executive was placed under house arrest and zero otherwise.
\item jailed. A dummy variable where one indicates the deposed executive was arrested, detained or jailed and zero otherwise.
\item tried. A dummy variable where one indicates the deposed executive was tried and zero otherwise.
\item fled. A dummy variable where one indicates the deposed executive fled the country and zero otherwise.
\item exile. A dummy variable where one indicates the deposed executive was banished from the country and zero otherwise.
}
}
\keyword{datasets}
