% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elections}
\alias{elections}
\title{US County Election Data}
\format{
A (tibble) data frame with 3,152 rows and 17 variables.
}
\source{
Mostly a port of socviz::county_data.
}
\usage{
data(elections)
}
\description{
County-level data with electoral results from 2012, 2016, and 2020. Note: some 2020 results based on estimates.
The variables are as follows:
\itemize{
\item fips. County FIPS ID.
\item name. Name of county.
\item state. State.
\item census_region. Region state belongs to according to Census.
\item pop. County population (2014).
\item female. Percent of county that is female (2013).
\item white. Percent of county that is white (2013).
\item black. Percent of county that is black (2013).
\item travel_time. Mean travel time to work (minutes), workers age 16+, 2009-2013.
\item land_area. County land area in square miles, 2010.
\item hh_income. Median household income, 2009-2013.
\item per_PARTY_2012. PARTY presidential vote, proportion, 2012.
\item per_PARTY_2016. PARTY presidential vote, proportion, 2016.
\item per_PARTY_2020. PARTY presidential vote, proportion, 2020.
}
}
\keyword{datasets}
