% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emission}
\alias{emission}
\title{Data on CO2 and Greenhouse Gas Emissions by Our World in Data}
\format{
A (tibble) data frame with 21,913 rows and 8 variables.
}
\source{
Our World in Data. https://github.com/owid/co2-data
}
\usage{
data(emission)
}
\description{
Country-year data from "Our World in Data" on CO2 and Greenhouse gas emissions.
The variables are as follows:
\itemize{
\item country. Name of country.
\item year. Year.
\item population. Population of country.
\item gdp. Gross domestic product (2011).
\item co2. Annual production-based emissions of carbon dioxide (CO2), measured in million tonnes. This is based on territorial emissions, which do not account for emissions embedded in traded goods.
\item coal_co2. Annual production-based emissions of carbon dioxide (CO2) from coal, measured in million tonnes. This is based on territorial emissions, which do not account for emissions embedded in traded goods.
\item cumulative_co2. Cumulative production-based emissions of carbon dioxide (CO2) since the first year of data availability, measured in million tonnes. This is based on territorial emissions, which do not account for emissions embedded in traded goods.
\item methane. Total methane emissions including land-use change and forestry, measured in million tonnes of carbon dioxide-equivalents.
}
}
\keyword{datasets}
