% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euro_elections}
\alias{euro_elections}
\title{PARLGOV Dataset on European elections}
\format{
A (tibble) data frame with 8,947 rows and 8 variables.
}
\source{
Döring, Holger, Constantin Huber, Philip Manow, Maike Hesse and Alexandra Quaas. 2023. Parliaments and governments database (ParlGov): Information on parties, elections and cabinets in established democracies.
}
\usage{
data(euro_elections)
}
\description{
Data on European election results, by political party.
The variables are as follows:
}
\details{
\itemize{
\item country. Country.
\item election. Type of election -- 'parliament' or 'ep' election.
\item date. Date the election took place. If the election lasted for several days, we code the last date.
\item party. Name of party.
\item party_id. Party ID.
\item vote_share. percentage of votes for a party.
\item seats. number of seats won by party. Seats coded for alliance members if information available. Coded as 0 for parties that won no seats and as missing value for electoral alliances if seats are coded at the level of alliance members.
\item total_seats. total number of seats.
}
}
\keyword{datasets}
