% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fatal}
\alias{fatal}
\title{iCasualties data on US fatalities in Iraq and Afghanistan}
\format{
A (tibble) data frame with 7,029 rows and 12 variables.
}
\source{
iCasualties.org
}
\usage{
data(fatal)
}
\description{
Data on American soldier fatalities during the US invasion of Iraq and Afghanistan, collected by iCasualties. Note: Data here may not reflect most updated numbers.
The variables are as follows:
\itemize{
\item date. Date of death.
\item name. Name of soldier.
\item rank. Soldier's rank.
\item branch. Branch of the army, navy, air force, or other entity that soldier was part ofc.
\item age. Age at time of death.
\item country. Country where soldier died.
\item province. Province where soldier died.
\item where. City or locale where soldier died.
\item source. Source of death (friendly vs hostile).
\item cause. Cause of death.
\item state. Home state of soldier.
\item city. Home city/town of soldier.
}
}
\keyword{datasets}
