% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fearon}
\alias{fearon}
\title{Fearon and Laitin (2003) replication data}
\format{
A (tibble) data frame with 6,610 rows and 10 variables.
}
\source{
From Fearon and Laitin (2003).
}
\usage{
data(fearon)
}
\description{
Replication data from the classic Fearon and Laitin (2003) study on the causes of civil wars.
The variables are as follows:
\itemize{
\item country. Name of country.
\item year. Year.
\item onset. Did country experience civil war onset this year? (1 = yes, 0 = no).
\item pop. Population, lagged one year.
\item polity2. Country's Polity2 score that year.
\item gdp_cap. GDP per capita, lagged one year, in thousands.
\item mtnest. Percent of country that is "mountainous".
\item oil. Does the country's oil exports exceed one third of export revenues? (1 = yes, 0 = no).
\item ef. Ethnic fractionalization index (the probability that two randomly drawn individuals in a country are from different ethnolinguistic group).
\item numlang.The number of distinct languages spoken by groups exceeding 1% the country's population.
}
}
\keyword{datasets}
