% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdp}
\alias{gdp}
\title{Economic and Population Panel Data for Latin America}
\format{
A (tibble) data frame with 4,013 rows and 11 variables.
}
\source{
The Varieties of Democracy (V-DEM) project, via: \url{https://github.com/xmarquez/vdem}
}
\usage{
data(gdp)
}
\description{
A country-year panel dataset from varied sources on economic and demographic indicators in Latin America.
The variables are as follows:
}
\details{
\itemize{
\item country. Name of country.
\item year. Year.
\item e_cow_exports. What is the total value of a country's exports? . Barbieri and Keshk (2016).
\item e_cow_imports. What is the total value of a country's imports? . Barbieri and Keshk (2016).
\item e_migdppc. What is the GDP per capita? . The Maddison Project Database (2018)
\item e_miinflat. What is the annual inflation rate? . Clio Infra (clio-infra.eu), based on Arroyo Abad, Davies and van Zanden (N.d.),
Montevideo- Oxford Latin America Economic History Database (http://moxlad.fcs.edu.uy/es/basededatos.html), De Zwart (2011a), De Zwart (2011b),
Reinhart and Rogoff (2011), Santing (N.d.), World Bank (2013).
\item e_total_oil_income_pc. What is the real value of a country's petroleum production? . Haber and Menaldo (2011).
\item e_mipopula. What is the total population (in thousands)? . Clio Infra (clio-infra.eu), drawing on Goldewijk, Beusen, Janssen (2010), History Database of Global Environment (www.pbl.nl/hyde).
\item e_miurbpop. What is the total urban population? . Clio Infra (clio-infra.eu) based on Bairoch, Batou, Chevre (1988), Chandler (1987), de Vries (1984),
History Database of Global Environment (www.pbl.nl/hyde), Etter, McAlpine, Possingham (2007), Grigg (1980), Goldewijk, Beusen, Janssen (2010),
Maddison (2001), Roz- man (1973), United Nations (2009a), United Nations (2009b), Urquhart and Buckley (1965), Van Zanden (2012).
\item e_peinfmor. What is the infant mortality rate? . Gapminder (gapminder.org), drawing on various sources.
\item e_pelifeex. What is the life expectancy? . Gapminder (gapminder.org), drawing on Human Mortality Database (2008),
Riley (2005a, 2005b), Human Life Table Database (2007), United Nations Population Division (2010). Clio Infra (clio-infra.eu),
drawing on Human Mortality Database (2008), Human Life Table Database (2017), World Bank (2013),
Montevideo-Oxford Latin America Economic History Database (http://moxlad.fcs.edu.uy/es/basededatos.html).
}
}
\keyword{datasets}
