% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ged}
\alias{ged}
\title{UCDP Geo-referenced Event Database}
\format{
A (tibble) data frame with 261,864 rows and 14 variables.
}
\source{
UCDP GED Version 2.11.
}
\usage{
data(ged)
}
\description{
A sample of the UCDP GED (version 21.1), information on violent events including casualty estimates.
The variables are as follows:
\itemize{
\item id. Unique event identifer.
\item year. Year event began.
\item month. Month event began.
\item day. Day event began.
\item country. Country of event.
\item type_of_violence. UCDP violence type.
\item conflict_name. Name of conflict.
\item side_a. Side A. See UCDP codebook.
\item side_b. Side B. See UCDP codebook.
\item longitude. Longitude.
\item latitude. Latitude.
\item low_death. Low estimate of number of deaths.
\item best_death. Best estimate of number of deaths.
\item high_death. High estimate of number of deaths.
}
}
\keyword{datasets}
