% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_arrests}
\alias{ice_arrests}
\title{ICE Administrative Arrests}
\format{
A (tibble) data frame with 362,202 rows and 17 variables.
}
\source{
Deportation Data Project. \url{https://deportationdata.org/}
}
\usage{
data(ice_arrests)
}
\description{
Data on ICE (Immigration and Customs Enforcement) administrative arrests, including information on apprehension circumstances, case status, and demographics. Likely duplicates have been removed.
The variables are as follows:
}
\details{
\itemize{
\item apprehension_date. Date of apprehension.
\item apprehension_state. State where apprehension took place.
\item apprehension_aor. ICE Area of Responsibility for the apprehension.
\item final_program. ICE program under which the arrest was made.
\item final_program_group. Broader program group.
\item apprehension_method. Method of apprehension (e.g., CAP Federal Incarceration, At-Large).
\item apprehension_criminality. Criminality category (1 = Convicted Criminal, 2 = Pending Criminal Charges, 3 = Other Immigration Violator).
\item case_status. Current case status (e.g., Deported/Removed, Active).
\item case_category. Detailed case category.
\item departed_date. Date of departure from the US (if applicable).
\item departure_country. Country of departure/removal.
\item final_order_yes_no. Whether a final order of removal has been issued (YES/NO).
\item final_order_date. Date of final order of removal.
\item birth_year. Year of birth.
\item citizenship_country. Country of citizenship.
\item gender. Gender.
\item apprehension_site_landmark. Landmark or location description of apprehension site.
}
}
\keyword{datasets}
