% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_detention}
\alias{ice_detention}
\title{ICE Detention Facility Daily Population}
\format{
A (tibble) data frame with 661,928 rows and 9 variables.
}
\source{
Deportation Data Project. \url{https://deportationdata.org/}
}
\usage{
data(ice_detention)
}
\description{
Daily population counts at ICE (Immigration and Customs Enforcement) detention facilities across the United States, broken down by gender and criminal status.
The variables are as follows:
}
\details{
\itemize{
\item detention_facility. Name of the detention facility.
\item state. State where the facility is located.
\item date. Date of the population count.
\item n_detained. Total number of individuals detained.
\item n_detained_at_midnight. Number of individuals detained at midnight.
\item n_detained_male. Number of male individuals detained.
\item n_detained_female. Number of female individuals detained.
\item n_detained_convicted_criminal. Number of detained individuals who are convicted criminals.
\item n_detained_possibly_under_18. Number of detained individuals who are possibly under 18 years old.
}
}
\keyword{datasets}
