% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invasions}
\alias{invasions}
\title{Data on Land Invasions in Brazil}
\format{
A (tibble) data frame with 4,062 rows and 7 variables.
}
\source{
Hidalgo, F. D., Naidu, S., Nichter, S., & Richardson, N. (2010).
Economic determinants of land invasions. The Review of Economics and Statistics, 92(3), 505-523.
}
\usage{
data(invasions)
}
\description{
Data on land invasions in Brazil, at the municipal level, from Hidalgo et. al (2010).
The variables are as follows:
\itemize{
\item ext_pov. Extreme Poverty (Percent), in 1991
\item landgini. Land Gini
\item occs. Land Occupations, aggregated over 1988-2004
\item logarea. Log (Land Area)
\item logfam. Log (Families in Land Occupations), aggregated over 1988-2004
\item logpop. Log (Population), in 1991
}
}
\keyword{datasets}
