% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leader}
\alias{leader}
\title{Data on state leaders, mostly from the Leader Experience, Attribute, and Decision (LEAD) data set.}
\format{
A (tibble) data frame with 17,686 rows and 16 variables.
}
\source{
`peacesciencer`.
}
\usage{
data(leader)
}
\description{
Leader-year data on biographical attributes of state leaders, including military experience, education, and health.
The variables are as follows:
\itemize{
\item country. Country.
\item gwcode. Gleditsch and Ward country code.
\item leader. Name of leader.
\item gender. Gender of leader.
\item year. Year.
\item yr_office. Years in office.
\item age. Age of leader.
\item edu. Educational attainment of leader.
\item mil_service. Does leader have military service? (1 = yes, 0 = no).
\item combat. Does leader have combat experience? (1 = yes, 0 = no).
\item rebel. Does leader have experience as part of a rebellion? (1 = yes, 0 = no).
\item yrs_exp. Previous years of experience in politics before becoming a leader.
\item phys_health. Does leader have physical health issues? (1 = yes, 0 = no).
\item mental_health. Does leader have mental health issues? (1 = yes, 0 = no).
\item will_force. Estimate of leader's willingness to use force.
\item will_force_sd. Standard deviation of estimate of leader's willingness to use force.
}
}
\keyword{datasets}
