% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\title{Movies data}
\format{
A (tibble) data frame with 4,139 rows and 10 variables.
}
\source{
IMDB via Kaggle.
}
\usage{
data(movies)
}
\description{
A sample of movies and their characteristics from IMDB (1950-2016).
The variables are as follows:
\itemize{
\item title. Name of the movie.
\item year. Year the movie was released.
\item decade. Decade the movie was released (factor).
\item director. Name of the movie's director.
\item genre1. The movie's first listed genre.
\item genre2. The movie's second listed genre.
\item duration. Duration of movie (in minutes).
\item gross. How much the movie grossed in theaters (in USD).
\item budget. How much the movie cost to make (in USD).
\item rating. The movie's IMDB rating (0-10).
}
}
\keyword{datasets}
