% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_shipname.R
\name{normalize_shipname}
\alias{normalize_shipname}
\title{Normalize shipname}
\usage{
normalize_shipname(name, ...)
}
\arguments{
\item{name}{A character string (or vector of strings) to be normalized}

\item{...}{Other args}
}
\value{
A normmalized shipname
}
\description{
Performs a series of string normalizations to get a clean shipname or callsign.
This was inspired by Jaeyoon Park's Python Module: \url{https://github.com/jaeyoonpark/shipdataprocess}.

\code{normalize_shipname} will do the following:
\itemize{
\item Make all letters into capital letters
\item Remove leading and trailing whitespaces
\item Remove linebreaks
\item Remove all references to vessel codes (\emph{i.e.} \code{MFV} stands for Marine Fishing Vessel)
\item Convert numbers expressed as letters to numbers (\emph{i.e.} \code{ONE} becomes \code{1})
\item Convert roman numerals to arabic numbers
\item Remove \code{NO.} when it references a number (\emph{i.e.} \code{BOAT NO.5} becomes \code{BOAT5})
\item Remove all special characters
}
}
\examples{
# Normalize a weird shipname
shipname <- "boatyMcB0atface <>$;!&"
normalize_shipname(shipname)
}
