% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parties}
\alias{parties}
\title{PARLGOV Dataset on European political parties}
\format{
A (tibble) data frame with 1,672 rows and 8 variables.
}
\source{
Döring, Holger, Constantin Huber, Philip Manow, Maike Hesse and Alexandra Quaas. 2023. Parliaments and governments database (ParlGov): Information on parties, elections and cabinets in established democracies.
}
\usage{
data(parties)
}
\description{
Data on the characteristics of European political parties. Note that I excluded parties where the family variable was uncoded.
The variables are as follows:
}
\details{
\itemize{
\item country. Country.
\item party. Name of party.
\item family. Broader ideological "family" (e.g., Conservative) that party belongs to.
\item left_right. 0--10 scale mean value in left/right dimension with data from Castles/Mair 1983, Huber/Inglehart 1995, Benoit/Laver 2006 and CHES 2010.
\item state_market. 0--10 scale mean value in 'regulation of the economy' dimension with data from Benoit/Laver 2006 and CHES 2010.
\item liberty_authority. 0--10 scale mean value in 'libertarian/authoritarian' dimension with data from Benoit/Laver 2006 and CHES 2010.
\item eu_anti_pro. 0--10 scale mean value in 'EU integration' dimension with data from Ray 1999, Benoit/Laver 2006 and CHES 2010.
\item party_id. Party ID.
}
}
\keyword{datasets}
