% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{Pokemon data}
\format{
A (tibble) data frame with 801 rows and 14 variables.
}
\source{
Kaggle.
}
\usage{
data(pokemon)
}
\description{
A dataset containing the characteristics of 800+ Pokemon.
The variables are as follows:
\itemize{
\item type1. Pokemon primary type.
\item type2. Pokemon secondary type.
\item height_m. Heigh in meters.
\item weight_kg. Weight in kilometers.
\item capture_rate. Measure of how easy it is to capture the Pokemon. Higher catch rates mean that the Pokémon is easier to catch, up to a maximum of 255.
\item hp. Determines how much damage a Pokémon can receive before fainting.
\item attack. Determines the power of a physical move.
\item defense. Determines how much damage a Pokemon receives when it is hit with a physical move.
\item sp_attack. Determines how much damage a Pokemon deals when using a special move.
\item sp_defense. Determines how much damage a Pokemon receives when it is hit with a special move.
\item speed. Determines the order of Pokémon that can act in battle.
\item generation. A grouping of the Pokémon games that separates them based on the Pokémon they include.
\item is_legendary. Whether the Pokemon is legendary (1 = yes, 0 = no).
}
}
\keyword{datasets}
