% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race}
\alias{race}
\title{Race and Colorism Survey Data from LAPOP}
\format{
A (tibble) data frame with 6,051 rows and 9 variables.
}
\source{
Latin American Public Opinion Project.
}
\usage{
data(race)
}
\description{
Survey data on race and colorism in Colombia, Brazil, Mexico, Guatemala,
and the Dominican Republic, from LAPOP, 2014 wave.
The variables are as follows:
\itemize{
\item pais. Country
\item q2. Age.
\item sex. Sex.
\item etid. Ethnicity.
\item ed. Years of schooling (0 to 18+ years)
\item colorr. Skin color (1 = lightest, 11 = darkest; see LAPOP color palette).
\item q10new_14. Monthly household income (0 = no income, 16 = highest income; see LAPOP).
}
}
\keyword{datasets}
