% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rebel_leader}
\alias{rebel_leader}
\title{Rebel Organization Leaders (ROLE) Database}
\format{
A (tibble) data frame with 425 rows and 20 variables.
}
\source{
Acosta, B., Huang, R., & Silverman, D. (2023). Introducing ROLE: A database of rebel leader attributes in armed conflict. Journal of Peace Research, 60(2), 352–361. https://doi.org/10.1177/00223433221077920
}
\usage{
data(rebel_leader)
}
\description{
Dataset that provides a wide range of biographical information on all top rebel, insurgent, and terrorist leaders who were active in civil wars between 1980 and 2011.
The variables are as follows:
}
\details{
\itemize{
\item leadercode. Leader code.
\item name. Full name.
\item state. The conflict country against which the leader fought.
\item conflict. Name of the conflict in which the leader fought.
\item group. Name of group.
\item gender. Gender of leader.
\item birth_year. Year of birth.
\item dead. Is leader dead? Coded as of December 31, 2018.
\item birth_place. Place of birth.
\item education. Highest level of completed education.
\item study_area. Area of study for highest degree achieved.
\item entry_method. How did the individual come to lead the organization?
\item western. Dummy variable where 1 means the leader obtained a degree (bachelor's, master's, doctorate, or equivalent) from a university in the US, UK, Austria, Belgium, Canada, Denmark, Finland, France, Germany, Greece, Ireland, Italy, the Netherlands, Norway, Portugal, Spain, Sweden, or Switzerland and 0 otherwise.
\item married. Dummy variable where 1 means the leader was married at some point and 0 otherwise.
\item children. Dummy variable where 1 means the leader had children and 0 otherwise.
\item religion. The religion with which the leader self-identified.
\item elite_family. Did the leader come from an elite family with political power or connections?
\item occupation. The leader's primary occupation prior to his/her becoming rebel organization leader.
\item train_abroad. Dummy variable where 1 means the leader received training abroad from a foreign military or foreign rebel group before becoming leader and 0 otherwise.
\item death_cause. Cause of death.
}
}
\keyword{datasets}
