% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{therm}
\alias{therm}
\title{US feeling thermometer data}
\format{
A (tibble) data frame with 4,989 rows and 21 variables.
}
\source{
July 2017 "Views of the Electorate" Research Survey.
}
\usage{
data(therm)
}
\description{
Feeling thermometer survey data on a variety of social groups in the USA. From from the July 2017 "Views of the Electorate" Research Survey.
The variables are as follows:
\itemize{
\item birth_year. Year respondent was born.
\item sex. Sex of respondent.
\item race. Race of respondent.
\item party_id. Respondent party identification (3 point).
\item educ. Highest level of educational attainment.
\item ft_GROUP. Feeling thermometer rating for GROUP (0 = feel very "cold" towards GROUP, 100 = feel very "warm" towards GROUP).
}
}
\keyword{datasets}
