% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade}
\alias{trade}
\title{Interstate Trade}
\format{
A (tibble) data frame with 4,263 rows and 13 variables.
}
\source{
From the `peacesciencer` package.
}
\usage{
data(trade)
}
\description{
State-year level Correlates of War data focusing on international trade, including imports, exports, and IGO membership.
The variables are as follows:
\itemize{
\item country. Name of country.
\item year. Year.
\item imports. Total imports of the state in current million USD.
\item exports. Total exports of the state in current million USD.
\item gdp. Estimated GDP in 2011 USD.
\item pop. Estimated population size.
\item land_borders. Number of countries bordered by land.
\item sea_borders. Number of countries bordered by sea.
\item min_cap_dist. Mininum distance to a foreign capital, in kilometers.
\item sum_igos. The sum of International Governmental Organizations for which the state is a member of any kind in a given year.
}
}
\keyword{datasets}
