% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdem}
\alias{vdem}
\title{Varieties of Democracy Indicators for Latin America}
\format{
A (tibble) data frame with 27,734 rows and 7 variables.
}
\source{
The Varieties of Democracy (V-DEM) project, via: https://github.com/vdeminstitute/vdemdata
}
\usage{
data(vdem)
}
\description{
A dataset from the Varieties of Democracy (V-DEM) project that characterizes
the different dimensions of democracy across Latin America.
The variables are as follows:
}
\details{
\itemize{
\item country. Country.
\item year. Year.
\item v2x_egaldem. To what extent is the ideal of egalitarian democracy achieved? . Interval, from low to high (0-1).
\item v2x_delibdem. To what extent is the ideal of deliberative democracy achieved? . Interval, from low to high (0-1).
\item v2x_partipdem. To what extent is the ideal of participatory democracy achieved? . Interval, from low to high (0-1).
\item v2x_libdem. To what extent is the ideal of liberal democracy achieved? . Interval, from low to high (0-1).
\item v2x_polyarchy. To what extent is the ideal of electoral democracy in its fullest sense achieved? . Interval, from low to high (0-1).
}
}
\keyword{datasets}
