% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vietnam}
\alias{vietnam}
\title{Dynamics of violence in civil war dataset}
\format{
A (tibble) data frame with 71,378 rows and 11 variables.
}
\source{
From Kalyvas and Kocher(2009).
}
\usage{
data(vietnam)
}
\description{
Replication data from Kalyvas and Kocher (2009) on the characteristics of the Hamlet Evaluation System during the Vietnam War.
The variables are as follows:
\itemize{
\item hamlet_id. 9-digit Hamlet Evaluation System identifier.
\item month. Month.
\item selective_terror. Did Vietcong use selective violence? (1 = yes, 0 = no).
\item control. Degree of control between state and rebels in the hamlet.
\item bombs. Number of bomb sorties in hamlet.
\item urban. Is hamlet mostly urban? (1 = yes, 0 = no).
\item pop. Hamlet population.
\item border. Distance to closest international border.
\item dev_index. Development index (cannot find details on what this means).
\item buddhist. Predominantly Buddhist hamlet? (1 = yes, 0 = no).
\item vietnamese. Predominantly Vietnamese hamlet? (1 = yes, 0 = no).
}
}
\keyword{datasets}
