% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wealth}
\alias{wealth}
\title{Data on Income and Household Assets in Honduras, 2018}
\format{
A (tibble) data frame with 1,560 rows and 20 variables.
}
\source{
Latin American Public Opinion Project. Honduras, 2018.
}
\usage{
data(wealth)
}
\description{
Survey data on income, household assets, and other markers of wealth/poverty in Honduras.
The variables are as follows:
\itemize{
\item r1. Television in Home
\item r3. Refrigerator in Home
\item r4. Landline in Home
\item r4a. Cellular Telephone in Home
\item r5. Number of Vehicles at the House
\item r6. Washing Machine in Home
\item r7. Microwave Oven in Home
\item r8. Owns Motorcycle
\item r12. Drinking water in Home
\item r14. Indoor Bathroom in Home
\item r15. Computer in Home
\item r16. Flat Panel TV in Home
\item r18. Internet Service in Home
\item ur. Urban/Rural
\item ed. Years of Schooling
\item q10new_18. Monthly Household Income
\item q14. Intends to Live or Work Abroad
\item fs2. Has Run Out of Food in the Last 3 Months
\item fs8. Has Gone without Meals in the Last 3 Months
\item wf1. Receives Government Assistance
}
}
\keyword{datasets}
