% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race}
\alias{race}
\title{Race and Colorism Survey Data from LAPOP}
\format{
A (tibble) data frame with 6,051 rows and 9 variables.
}
\source{
Latin American Public Opinion Project.
}
\usage{
data(race)
}
\description{
Survey data on race and colorism in Colombia, Brazil, Mexico, Guatemala,
and the Dominican Republic, from LAPOP, 2014 wave.
The variables are as follows:
\itemize{
\item pais. Country
\item q2. Age.
\item etid. Ethnicity.
\item ed. Years of schooling.
\item colorr. Skin color (1 = lightest, 11 = darkest; see LAPOP color palette).
\item q10new_14. Monthly household income (0 = no income, 16 = highest income; see labels).
\item d5. Approval of homosexuals' right to run for office (1 = strongly disapprove, 10 = strongly approve).
\item d6. Approval of same-sex couples' right to marry (1 = strongly disapprove, 10 = strongly approve).
}
}
\keyword{datasets}
